requirements.txt
Django==1.3.1
Fabric==1.4.2
MySQL-python==1.2.3
Pillow==1.7.7
South==0.7.3
Werkzeug==0.8.3
boto==2.3.0
commonware==0.2.4
distribute==0.6.24
-e git+git@github.com:Jenso/django-braces.git@608c8704e49bb40f1610213f58621a3fdd26bf5c#egg=django_braces-dev
django-debug-toolbar==0.9.4
django-defaultsite==1.1
django-extensions==0.9
django-liveserver==0.1a-2
-e git+git@github.com:Jenso/django-registration@346823d4df5db902dffe1ec2f447d38a52889c47#egg=django_registration-dev
django-ses==0.4.1
django-social-auth==0.6.6
django-storages==1.1.4
django-su==0.2
django-tastypie==0.9.11
google-api-python-client==1.0c2
gunicorn==0.14.2
httplib2==0.7.2
ipdb==0.6.1
iptools==0.4.0
ipython==0.12
mimeparse==0.1.3
mock==0.8.0
oauth2==1.5.211
pep8==1.2
pycrypto==2.6
python-dateutil==1.5
python-gflags==2.0
python-openid==2.2.5
raven==1.7.5
selenium==2.22.1
simplejson==2.5.2
sorl-thumbnail==11.12
ssh==1.7.14
unittest2==0.5.1
wsgiref==0.1.2
